headers: start moving relevant types to top level
